Fix a bug in large_sub()

+++ large_int.c
@@ -149,7 +149,6 @@ static void large_add(LargeInt* former, LargeInt* latter, LargeInt* result) {
 
 // former - latterを行う
 // 符号は全く気にしない
-// FIXME: 0x1ffffffff - 0x1 => 0x1ffffffff
 static void large_sub(LargeInt* former, LargeInt* latter, LargeInt* result) {
     // large_sub(a, b, a)などにも対応するためresultは最後に触る
     LargeInt buffer;
@@ -162,7 +161,7 @@ static void large_sub(LargeInt* former, LargeInt* latter, LargeInt* result) {
         // 片方のリストが短くてもぬるぽしないようにsecurely~を使う
         // 繰り下がりのぶんを予め足しておく
         unsigned long new_value =
-            (1 << (kHexDigitsInUInt * 4)) +
+            ((unsigned long)1 << (kHexDigitsInUInt * 4)) +
             (unsigned long)securely_get_value(former_node) -
             (unsigned long)securely_get_value(latter_node) - carry;
         // new_valueの下半分を取り出して代入
